refactor(map): drop Next suffix from migrated components
The legacy Leaflet versions are gone, so the *Next-suffixed names are
just dead semantic baggage now. Renames:

- $components/CommunityRailNext.svelte → CommunityRail.svelte
- $components/MultiPlaceMapNext.svelte → MultiPlaceMap.svelte
- $components/area/AreaMapNext.svelte → AreaMap.svelte

Plus consumer updates in src/routes/map/+page.svelte,
src/routes/user/saved/+page.svelte, src/components/area/AreaPage.svelte,
and one internal log-string in CommunityRail.svelte.
