+++ src/components/MultiPlaceMap.svelte
