+++ src/components/area/AreaMap.svelte
